Algorithm working with boundary conditions.
Renamed the files tdma.cpp and tdma.h that use the algorithm to
gaussianElimination.cpp and gaussianElimination.h to better reflect what
they do.

Finally made the boundary conditions right, and the algorithm is now
very close (pretty much the same) to the analytical solution, even for "low" N.
Made the indexing so that it better represents what actually happens with the
elements in the tridiagonal matrix. This helped me get the boundary
conditions right, and made the algorithm work like it should.